Changelog: 1. removed forward slash in front of all URL redirects of following pages for hosting reasons  -

+++ app/LearnHub/LearnHub/AppCode/dao/Course_elearnDAO.cs
@@ -540,7 +540,7 @@ namespace LearnHub.AppCode.dao
         }
 
 
-        public void updateCourse(int courseID, int categoryID, string name, string description, int hours, DateTime start, DateTime expiry) // Update.
+        public void updateCourse(int courseID, int categoryID, string name, string description, double hours, DateTime start, DateTime expiry) // Update.
         {
             SqlConnection conn = new SqlConnection();
